backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
websockets==12.0

# Testing dependencies
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0 